@@ -1,35 +1,27 @@
-import { NgModule } from '@angular/core';
-import { BrowserModule } from '@angular/platform-browser';
-
-import { AppComponent } from './app.component';
-import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
-import {HttpClientModule} from '@angular/common/http';
+import {NgModule} from '@angular/core';
+import {BrowserModule} from '@angular/platform-browser';
+import {AppComponent} from './app.component';
+import {BrowserAnimationsModule} from '@angular/platform-browser/animations';
+import {provideHttpClient, withInterceptorsFromDi} from '@angular/common/http';
 import {SharedModule} from '../shared/shared.module';
-import {Router, RouterModule, Scroll, Event} from '@angular/router';
+import {RouterModule} from '@angular/router';
 import {AppRoutingModule} from './app-routing.module';
 import {MatButtonModule} from '@angular/material/button';
 import {MatIconModule} from '@angular/material/icon';
 import {MatSidenavModule} from '@angular/material/sidenav';
-import {ViewportScroller} from '@angular/common';
-import {filter} from 'rxjs/operators';
 
 @NgModule({
   declarations: [
     AppComponent
   ],
-  imports: [
-    BrowserModule,
+  bootstrap: [AppComponent], imports: [BrowserModule,
     AppRoutingModule,
     BrowserAnimationsModule,
-    HttpClientModule,
     SharedModule,
     RouterModule,
     MatButtonModule,
     MatIconModule,
-    MatSidenavModule
-  ],
-  providers: [],
-  bootstrap: [AppComponent]
+    MatSidenavModule], providers: [provideHttpClient(withInterceptorsFromDi())]
 })
 export class AppModule {
 }
