@@ -1,11 +1,11 @@
 import {Component, OnInit} from '@angular/core';
-import {icons} from '../../../../shared/data/icons';
 import {links} from '../../../../shared/data/links';
 
 @Component({
-  selector: 'app-home-page',
-  templateUrl: './home-page.component.html',
-  styleUrls: ['./home-page.component.scss']
+    selector: 'app-home-page',
+    templateUrl: './home-page.component.html',
+    styleUrls: ['./home-page.component.scss'],
+    standalone: false
 })
 export class HomePageComponent implements OnInit {
 
