@@ -2,9 +2,10 @@ import {Component, Input, OnInit} from '@angular/core';
 import {Project} from '../../../../shared/models/project';
 
 @Component({
-  selector: 'app-project-card',
-  templateUrl: './project-card.component.html',
-  styleUrls: ['./project-card.component.scss']
+    selector: 'app-project-card',
+    templateUrl: './project-card.component.html',
+    styleUrls: ['./project-card.component.scss'],
+    standalone: false
 })
 export class ProjectCardComponent implements OnInit {
 
