@@ -3,9 +3,10 @@ import {myMainProjects} from '../../../../shared/data/projects';
 import {MainProject} from '../../../../shared/models/main-project';
 
 @Component({
-  selector: 'app-projects-page',
-  templateUrl: './projects-page.component.html',
-  styleUrls: ['./projects-page.component.scss']
+    selector: 'app-projects-page',
+    templateUrl: './projects-page.component.html',
+    styleUrls: ['./projects-page.component.scss'],
+    standalone: false
 })
 export class ProjectsPageComponent implements OnInit {
 
