@@ -1,9 +1,10 @@
 import { Component, OnInit } from '@angular/core';
 
 @Component({
-  selector: 'app-resume-page',
-  templateUrl: './resume-page.component.html',
-  styleUrls: ['./resume-page.component.scss']
+    selector: 'app-resume-page',
+    templateUrl: './resume-page.component.html',
+    styleUrls: ['./resume-page.component.scss'],
+    standalone: false
 })
 export class ResumePageComponent implements OnInit {
 
