@@ -1,18 +1,19 @@
-import {Component, Input, OnInit} from '@angular/core';
-import {ThemeService} from '../../services/theme-service';
+import {Component, Input} from '@angular/core';
+import {ThemeModeService} from '../../services/theme-mode-service';
 
 @Component({
-  selector: 'app-basic-toolbar',
-  templateUrl: './basic-toolbar.component.html',
-  styleUrls: ['./basic-toolbar.component.scss']
+    selector: 'app-basic-toolbar',
+    templateUrl: './basic-toolbar.component.html',
+    styleUrls: ['./basic-toolbar.component.scss'],
+    standalone: false
 })
 export class BasicToolbarComponent {
 
   @Input() showThemeToggle = true;
 
-  isDarkTheme$ = this.themeService.isDarkTheme;
+  appTheme$ = this.themeService.themeMode$;
 
-  constructor(private themeService: ThemeService) {
+  constructor(private themeService: ThemeModeService) {
   }
 
 }
