@@ -1,7 +1,7 @@
 import {Link} from '../models/link';
-import {icons} from './icons';
+import {registeredIcons} from './registeredIcons';
 
 export const links: { [name: string]: Link } = {
-  github: {icon: icons.github, url: 'https://github.com/kmartin0'},
-  linkedIn: {icon: icons.linkedIn, url: 'https://linkedin.com/in/pk-martin'}
+  github: {icon: registeredIcons.github, url: 'https://github.com/kmartin0'},
+  linkedIn: {icon: registeredIcons.linkedIn, url: 'https://linkedin.com/in/pk-martin'}
 }
