@@ -1,5 +1,5 @@
 import {Project} from "../../models/project";
-import {icons} from "../icons";
+import {registeredIcons} from "../registeredIcons";
 import {MainProject} from "../../models/main-project";
 
 export const projectDartsMatcherWeb: Project = {
@@ -9,23 +9,23 @@ export const projectDartsMatcherWeb: Project = {
   features: ['Light/dark mode.', 'Responsive design for mobile and desktop.', 'Login and register.', 'Create and play darts matches with an online scoreboard with up to 4 players.', 'Configure and play against a darts bot.', 'Recover from wrong score inputs.', 'View match information, statistics and a recap.'],
   technical: [],
   imageUrls: [
-    {name: 'Homepage', url: 'https://github.com/kmartin0/assets/blob/master/darts-matcher-web/darts_matcher_web_login_register.png?raw=true'},
-    {name: 'Dashboard Friend', url: 'https://github.com/kmartin0/assets/blob/master/darts-matcher-web/darts_matcher_web_friends.png?raw=true'},
-    {name: 'Create Match', url: 'https://github.com/kmartin0/assets/blob/master/darts-matcher-web/darts_matcher_web_create_match.png?raw=true'},
-    {name: 'Create Match Add Player', url: 'https://github.com/kmartin0/assets/blob/master/darts-matcher-web/darts_matcher_web_add_player.png?raw=true'},
-    {name: 'Create Match Add Bot', url: 'https://github.com/kmartin0/assets/blob/master/darts-matcher-web/darts_matcher_web_add_bot.png?raw=true'},
-    {name: 'Match Lobby', url: 'https://github.com/kmartin0/assets/blob/master/darts-matcher-web/darts_matcher_web_lobby.png?raw=true'},
-    {name: 'Match Scoreboard Dark', url: 'https://github.com/kmartin0/assets/blob/master/darts-matcher-web/darts_matcher_web_scoreboard_dark.png?raw=true'},
-    {name: 'Match Scoreboard Light', url: 'https://github.com/kmartin0/assets/blob/master/darts-matcher-web/darts_matcher_web_scoreboard_light.png?raw=true'},
-    {name: 'Match Information', url: 'https://github.com/kmartin0/assets/blob/master/darts-matcher-web/darts_matcher_web_match_information.png?raw=true'},
-    {name: 'Match Statistics', url: 'https://github.com/kmartin0/assets/blob/master/darts-matcher-web/darts_matcher_web_match_statistics.png?raw=true'},
-    {name: 'Match Recap', url: 'https://github.com/kmartin0/assets/blob/master/darts-matcher-web/darts_matcher_web_match_recap.png?raw=true'},
+    {name: 'Homepage', url: 'https://raw.githubusercontent.com/kmartin0/assets/refs/heads/master/darts-matcher-web/darts_matcher_web_login_register.png'},
+    {name: 'Dashboard Friend', url: 'https://raw.githubusercontent.com/kmartin0/assets/refs/heads/master/darts-matcher-web/darts_matcher_web_friends.png'},
+    {name: 'Create Match', url: 'https://raw.githubusercontent.com/kmartin0/assets/refs/heads/master/darts-matcher-web/darts_matcher_web_create_match.png'},
+    {name: 'Create Match Add Player', url: 'https://raw.githubusercontent.com/kmartin0/assets/refs/heads/master/darts-matcher-web/darts_matcher_web_add_player.png'},
+    {name: 'Create Match Add Bot', url: 'https://raw.githubusercontent.com/kmartin0/assets/refs/heads/master/darts-matcher-web/darts_matcher_web_add_bot.png'},
+    {name: 'Match Lobby', url: 'https://raw.githubusercontent.com/kmartin0/assets/refs/heads/master/darts-matcher-web/darts_matcher_web_lobby.png'},
+    {name: 'Match Scoreboard Dark', url: 'https://raw.githubusercontent.com/kmartin0/assets/refs/heads/master/darts-matcher-web/darts_matcher_web_scoreboard_dark.png'},
+    {name: 'Match Scoreboard Light', url: 'https://raw.githubusercontent.com/kmartin0/assets/refs/heads/master/darts-matcher-web/darts_matcher_web_scoreboard_light.png'},
+    {name: 'Match Information', url: 'https://raw.githubusercontent.com/kmartin0/assets/refs/heads/master/darts-matcher-web/darts_matcher_web_match_information.png'},
+    {name: 'Match Statistics', url: 'https://raw.githubusercontent.com/kmartin0/assets/refs/heads/master/darts-matcher-web/darts_matcher_web_match_statistics.png'},
+    {name: 'Match Recap', url: 'https://raw.githubusercontent.com/kmartin0/assets/refs/heads/master/darts-matcher-web/darts_matcher_web_match_recap.png'},
   ],
-  stockImageUrl: {name: 'Match Scoreboard Light', url: 'https://github.com/kmartin0/assets/blob/master/darts-matcher-web/darts_matcher_web_scoreboard_light.png?raw=true'},
-  techs: [icons.web, icons.typescript, icons.angular],
+  stockImageUrl: {name: 'Match Scoreboard Light', url: 'https://raw.githubusercontent.com/kmartin0/assets/refs/heads/master/darts-matcher-web/darts_matcher_web_scoreboard_light.png'},
+  techs: [registeredIcons.web, registeredIcons.typescript, registeredIcons.angular],
   links: [
-    {icon: icons.github, url: 'https://github.com/kmartin0/darts-matcher-web'},
-    {icon: icons.hostSelf, url: 'https://dartsmatcher.kmartin.nl'},
+    {icon: registeredIcons.github, url: 'https://github.com/kmartin0/darts-matcher-web'},
+    {icon: registeredIcons.hostSelf, url: 'https://dartsmatcher.kmartin.nl'},
   ]
 }
 
@@ -37,10 +37,10 @@ export const projectDartsMatcherApi: Project = {
   technical: ['Layered architecture.', 'MongoDB.', 'Globalized error handling.', 'Localization.', 'OAuth 2 and JWT for user authentication/authorization (Spring Security).', 'Websockets'],
   imageUrls: [],
   stockImageUrl: {name: 'terminal', url: 'assets/terminal.png'},
-  techs: [icons.api, icons.java, icons.springBoot, icons.websocket, icons.mongodb],
+  techs: [registeredIcons.api, registeredIcons.java, registeredIcons.springBoot, registeredIcons.websocket, registeredIcons.mongodb],
   links: [
-    {icon: icons.github, url: 'https://github.com/kmartin0/darts-matcher-api'},
-    {icon: icons.hostSelf, url: 'https://api.dartsmatcher.kmartin.nl'}
+    {icon: registeredIcons.github, url: 'https://github.com/kmartin0/darts-matcher-api'},
+    {icon: registeredIcons.hostSelf, url: 'https://api.dartsmatcher.kmartin.nl'}
   ]
 }
 
