@@ -4,11 +4,14 @@ import {BasicToolbarComponent} from './components/basic-toolbar/basic-toolbar.co
 import {ThemeToggleComponent} from './components/theme-toggle/theme-toggle.component';
 import {LoseFocusAfterClickDirective} from './directives/lose-focus-after-click-directive/lose-focus-after-click.directive';
 import {MatToolbarModule} from '@angular/material/toolbar';
-import {MatButtonModule} from '@angular/material/button';
 import {MatIconModule} from '@angular/material/icon';
-import {MatTooltipModule} from '@angular/material/tooltip';
 import { PageNotFoundComponent } from './components/page-not-found/page-not-found.component';
 import {SwiperModule} from 'swiper/angular';
+import {MatButtonModule} from "@angular/material/button";
+import {MatTooltipModule} from "@angular/material/tooltip";
+import {MatMenu, MatMenuItem, MatMenuTrigger} from "@angular/material/menu";
+import {MatRadioButton, MatRadioGroup} from "@angular/material/radio";
+import {MatRipple} from "@angular/material/core";
 
 
 @NgModule({
@@ -24,7 +27,13 @@ import {SwiperModule} from 'swiper/angular';
     MatButtonModule,
     MatIconModule,
     MatTooltipModule,
-    SwiperModule
+    SwiperModule,
+    MatMenu,
+    MatMenuTrigger,
+    MatMenuItem,
+    MatRadioGroup,
+    MatRadioButton,
+    MatRipple
   ],
   exports: [
     BasicToolbarComponent,
